Update : Responsive & Navbar fix
