+++ src/index.css
@@ -14,10 +14,10 @@
     @apply text-[22px] md:text-[26px] lg:text-[34px] font-medium;
   }
   h2 {
-    @apply text-[20px] md:text-[22px] lg:text-[26px] font-medium;
+    @apply text-[18px] md:text-[22px] lg:text-[26px] font-medium;
   }
   h3 {
-    @apply text-[18px] lg:text-[22px] font-medium;
+    @apply text-[16px] lg:text-[22px] font-medium;
   }
 }
