+++ src/layouts/cart/EntireDetail .jsx
@@ -1,11 +1,13 @@
 import React from "react";
+
+/* ================== Import layouts from same level ================== */
 import ToolTipHeader from "./ToolTipHeader";
 import BookingDetail from "./BookingDetail";
 import ReservePrice from "./ReservePrice";
 
-const EntireDetail  = () => {
+const EntireDetail = () => {
   return (
-    <div className="p-12 bg-white mx-[1.5em] w-[calc(100%-3em)] relative -translate-y-32 min-h-[40em] shadow-xl rounded-xl">
+    <div className="p-4 md:p-12 bg-white mx-[0.5em] md:mx-[1.5em] w-[calc(100%-1em)] md:w-[calc(100%-3em)] relative -translate-y-32 min-h-[40em] shadow-xl rounded-xl">
       <h3 className="font-normal text-sky-600">ราคาขาย</h3>
       <h2 className="font-semibold text-bluegrey-dark-2 leading-loose">
         THB 32,050,000
@@ -36,4 +38,4 @@ const EntireDetail  = () => {
   );
 };
 
-export default EntireDetail ;
+export default EntireDetail;
