+++ src/layouts/navbar/BrandIcon.jsx
@@ -1,0 +1,9 @@
+import { Link } from "react-router-dom";
+const BrandIcon = () => (
+  <Link to="/" className="text-lg lg:text-2xl font-medium text-[#0D47A1]">
+    Nikar
+    <span className="text-[#263238]">.com</span>
+  </Link>
+);
+
+export default BrandIcon;
